api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
